[framework] refactor: colocate runner protocol with framework base
